+++ src/components/emp/EmpTopMenu.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+
+export default function EmpTopMenu() {
+    return (
+        <>
+            <div>EmpTopMenu</div>
+            <div className='top-menu-grid'>
+                <div>상담 관리</div>
+                <div>내 정보 관리</div>
+                <div>지도교수 관리</div>
+            </div>
+        </>
+    )
+}
